+++ block2/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Added
+* Implemented `Debug` for `Block`, `ConcreteBlock`, `RcBlock` and
+  `GlobalBlock`.
+
 
 ## 0.2.0-alpha.4 - 2022-06-13